CASSANDRASC-55: Extract the in-jvm dtest template for use in other projects.
This also required moving some additional classes to `common` so they
could be used in the new test-framework project.
