+++ .gitignore
@@ -77,7 +77,7 @@ src/dist/*
 *.logdir_IS_UNDEFINED
 
 # Sidecar version - generated by build.gradle
-src/main/resources/sidecar.version
+common/src/main/resources/sidecar.version
 
 # Sidecar copyDist files copied to root directory
 agents
