+++ common/src/main/java/org/apache/cassandra/sidecar/cluster/instance/InstanceMetadata.java
