+++ settings.gradle
@@ -22,6 +22,7 @@ rootProject.name = "cassandra-sidecar"
 include "adapters:base"
 include "common"
 include "client"
+include "test-framework"
 include "vertx-client"
 include "vertx-client-shaded"
 include "docs"
