+++ src/test/integration/org/apache/cassandra/sidecar/common/JmxClientTest.java
@@ -23,9 +23,9 @@ import java.util.Map;
 
 import org.apache.cassandra.distributed.api.IInstanceConfig;
 import org.apache.cassandra.distributed.api.IUpgradeableInstance;
-import org.apache.cassandra.sidecar.common.testing.CassandraIntegrationTest;
-import org.apache.cassandra.sidecar.common.testing.CassandraTestContext;
 import org.apache.cassandra.sidecar.common.utils.GossipInfoParser;
+import org.apache.cassandra.sidecar.testing.CassandraIntegrationTest;
+import org.apache.cassandra.sidecar.testing.CassandraTestContext;
 
 import static org.assertj.core.api.Assertions.assertThat;
